fix(p5): beyond-L10 batch 3/10 — admin/*.php top-level (part 1)
Same mechanical (bool)-cast / useless-cast-removal fix as batches 1-2,
19 top-level admin/*.php files this time (admin.php, albums.php,
batch_manager*.php, cat_*.php, configuration.php, element_set_ranks.php,
extend_for_templates.php, group_*.php, intro.php, languages_new.php,
maintenance_*.php, menubar.php).

One extra @phpstan-ignore in admin/configuration.php:
order_by_is_local() already has 2 documented ignores (isset.offset,
logicalOr.alwaysFalse) explaining PHPStan can't see the optional
local/config/config.inc.php file's content; the new return.tooWideBool
finding is the same tension, so it's suppressed with the same
rationale rather than narrowing the return type to a misleading `false`.

Loose-comparison/empty()/strict-function-call and type-coverage
findings in these files are intentionally deferred, same as prior batches.

+++ admin/menubar.php
@@ -98,7 +98,7 @@ foreach ($reg_blocks as $id => $block) {
 if (isset($_POST['submit']) and is_webmaster()) {
     foreach ($mb_conf as $id => $pos) {
         $hide = isset($_POST['hide_' . $id]);
-        $mb_conf[$id] = ($hide ? -1 : +1) * (int) abs($pos);
+        $mb_conf[$id] = ($hide ? -1 : +1) * abs($pos);
 
         $pos_input = $_POST['pos_' . $id] ?? null;
         $pos = is_numeric($pos_input) ? (int) $pos_input : 0;
